requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
Pillow==10.1.0
fpdf==1.7.2
requests==2.31.0